refactor: sorting files to the correct folder
