+++ lib/pages/auth/register_cred_page.dart
@@ -1,11 +1,11 @@
 import 'package:smart_doku/pages/auth/register_pages.dart';
 import 'package:flutter/material.dart';
 import 'package:dropdown_button2/dropdown_button2.dart';
-import 'package:smart_doku/utils/formatter.dart';
-import 'package:smart_doku/utils/hover.dart';
-import 'package:smart_doku/utils/dialog.dart';
-import 'package:smart_doku/utils/function.dart';
-import 'package:smart_doku/utils/map.dart';
+import 'package:smart_doku/utils/handlers/formatter.dart';
+import 'package:smart_doku/utils/helper/hover.dart';
+import 'package:smart_doku/utils/handlers/dialog.dart';
+import 'package:smart_doku/utils/handlers/function.dart';
+import 'package:smart_doku/utils/helper/map.dart';
 import 'dart:ui';
 import 'dart:math';
 import 'package:smart_doku/utils/platform/platform_info.dart';
