+++ lib/pages/forms/users/detail_keluar_page.dart
@@ -3,8 +3,8 @@ import 'dart:ui';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:smart_doku/models/surat.dart';
 import 'package:smart_doku/utils/handlers/dateparser.dart';
-import 'package:smart_doku/utils/widget.dart';
-import 'package:smart_doku/utils/function.dart';
+import 'package:smart_doku/utils/widget/widget.dart';
+import 'package:smart_doku/utils/handlers/function.dart';
 
 class DetailPageKeluar extends StatefulWidget {
   final SuratKeluarModel? suratData; // Parameter untuk data surat
