+++ lib/pages/views/admins/desktop/home_page_admin_desktop.dart
@@ -4,7 +4,7 @@ import 'package:line_icons/line_icons.dart';
 import 'package:smart_doku/services/surat.dart';
 import 'package:smart_doku/services/user.dart';
 import 'dart:ui';
-import 'package:smart_doku/utils/function.dart';
+import 'package:smart_doku/utils/handlers/function.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class AdminDashboard extends StatefulWidget {
