+++ lib/pages/views/admins/phones/surat_disposisi_page_admin.dart
@@ -3,10 +3,10 @@ import 'package:package_info_plus/package_info_plus.dart';
 import 'package:smart_doku/models/surat.dart';
 import 'package:smart_doku/services/surat.dart';
 import 'dart:ui';
-import 'package:smart_doku/utils/dialog.dart';
-import 'package:smart_doku/utils/function.dart';
+import 'package:smart_doku/utils/handlers/dialog.dart';
+import 'package:smart_doku/utils/handlers/function.dart';
 import 'package:smart_doku/utils/handlers/dateparser.dart';
-import 'package:smart_doku/utils/widget.dart';
+import 'package:smart_doku/utils/widget/widget.dart';
 
 class DispositionLetterPageAdmin extends StatefulWidget {
   final Map<String, dynamic>? suratData;
