+++ lib/pages/views/admins/phones/surat_permohonan_page_admin.dart
@@ -6,10 +6,10 @@ import 'package:smart_doku/services/user.dart';
 import 'package:smart_doku/utils/refreshList/Mobile_Refresh_List.dart';
 import 'dart:ui';
 import 'dart:io';
-import 'package:smart_doku/utils/dialog.dart';
-import 'package:smart_doku/utils/function.dart';
+import 'package:smart_doku/utils/handlers/dialog.dart';
+import 'package:smart_doku/utils/handlers/function.dart';
 import 'package:smart_doku/utils/handlers/dateparser.dart';
-import 'package:smart_doku/utils/map.dart';
+import 'package:smart_doku/utils/helper/map.dart';
 import 'package:smart_doku/utils/search/SuratMasukFunction.dart';
 import 'package:smart_doku/utils/card/surat_masuk_card_mobile.dart';
