+++ lib/pages/views/users/desktop/profile_user_page.dart
@@ -4,9 +4,9 @@ import 'package:line_icons/line_icons.dart';
 import 'package:smart_doku/models/user.dart';
 import 'package:smart_doku/services/user.dart';
 import 'dart:ui';
-import 'package:smart_doku/utils/function.dart';
+import 'package:smart_doku/utils/handlers/function.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:smart_doku/utils/widget.dart';
+import 'package:smart_doku/utils/widget/widget.dart';
 import 'package:smart_doku/models/surat.dart';
 
 class UserProfile extends StatefulWidget {
