+++ lib/pages/views/users/desktop/surat_keluar_page_desktop.dart
@@ -6,11 +6,11 @@ import 'package:smart_doku/services/surat.dart';
 import 'package:smart_doku/services/user.dart';
 import 'dart:ui';
 import 'dart:io';
-import 'package:smart_doku/utils/dialog.dart';
-import 'package:smart_doku/utils/function.dart';
+import 'package:smart_doku/utils/handlers/dialog.dart';
+import 'package:smart_doku/utils/handlers/function.dart';
 import 'package:smart_doku/utils/handlers/dateparser.dart';
-import 'package:smart_doku/utils/map.dart';
-import 'package:smart_doku/utils/widget.dart';
+import 'package:smart_doku/utils/helper/map.dart';
+import 'package:smart_doku/utils/widget/widget.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class OutgoingLetterPageDesktop extends StatefulWidget {
